+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## vNext
 
+- Delay Funnel's DNS configuration test (new secretfile sub-option)
 - Refactor caching
 
 ## 0.6.4